+++ CHANGELOG.md
@@ -146,6 +146,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Package build (#238).
+- Allow any MessagePack supported type as a request key (#240).
 
 ## 0.9.0 - 2022-06-20